assignments/assignment1/requirements.txt
streamlit
transformers
torch
Pillow
sentencepiece
datasets
scikit-learn
matplotlib
seaborn
numpy
pandas
PyYAML
python-dotenv
wandb
